requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
gevent==1.4.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud==0.34.0
google-cloud-pubsub==1.2.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
